refactor: User 레이어 수정
