+++ src/main/java/kr/hhplus/be/server/user/infrastructure/PointHistoryRepository.java
@@ -1,5 +1,6 @@
-package kr.hhplus.be.server.user;
+package kr.hhplus.be.server.user.infrastructure;
 
+import kr.hhplus.be.server.user.domain.PointHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PointHistoryRepository extends JpaRepository<PointHistory, Long> {
