+++ src/main/java/kr/hhplus/be/server/user/infrastructure/UserRepository.java
@@ -1,5 +1,6 @@
-package kr.hhplus.be.server.user;
+package kr.hhplus.be.server.user.infrastructure;
 
+import kr.hhplus.be.server.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository  extends JpaRepository<User, Long> {
