+++ src/main/java/kr/hhplus/be/server/user/presentation/UserController.java
@@ -1,6 +1,10 @@
-package kr.hhplus.be.server.user;
+package kr.hhplus.be.server.user.presentation;
 
 import kr.hhplus.be.server.common.exception.EntityNotFoundException;
+import kr.hhplus.be.server.user.application.UserService;
+import kr.hhplus.be.server.user.domain.User;
+import kr.hhplus.be.server.user.dto.response.PointBalanceResponse;
+import kr.hhplus.be.server.user.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
